triageai-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
pydantic==2.5.3
pydantic-settings==2.1.0
xgboost==2.0.3
scikit-learn==1.4.0
shap==0.44.0
pandas==2.1.4
numpy==1.26.3
google-generativeai==0.3.2
python-multipart==0.0.6
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
